requirements.txt
python-telegram-bot==20.0
gtts
langdetect
requests
python-dotenv
